Deal with the edge condition when input is INT_MIN
Suppose we take absolute of signed integer x, abs(x)
when x is INT_MIN, the result is undefined,
but when x is INT_MIN, the absolute value of x is simply equals to
(unsigned) x,
so we take this property, and carefully transfer x to its absolute
value.

With this patch, the runtime error below is gone.

divid_int.c:45:9: runtime error: negation of -2147483648 cannot be represented in type 'int'; cast to an unsigned type to negate this value to itself