Use channel to notify signal
This commit practice using signal.Notify() to register the signal
we want to handle, e.g. syscall.SIGINT, syscall.SIGTERM, and start
a go routine to handle these signal.
When the signal is triggered, we print the signal name,
and send true using another channel 'done' to
tell main goroutine to exit noremally.

Ref:
    - [Go by Example: Signals](https://gobyexample.com/signals)