LGWT: Select
This program take two url, and make GET request to them, and
compare the time spent, return the faster one.
In the test, we use httptest.NewServer() and make GET request to
httptest.NewServer().URL, instead of the real URL.
It makes the test self-contained, not affected by status of real-world
internet connection traffic.
And after testing that, we use Close() method to shutdown our test server.

Ref:
    - [LGWT - Select](https://quii.gitbook.io/learn-go-with-tests/go-fundamentals/select)