Seek problem of file reading
In the test, when we Decode the string from strings.Reader to go data structure,
the position of reading is pointed to the end, so next time we call
GetLeague() again, we can't get any data.

We need a solution that can modify the read position, let it
point to starting point every time.

Ref:
    - [LGWT - IO, sorting](https://quii.gitbook.io/learn-go-with-tests/build-an-application/io#seeking-problems)