Add comment to SpyTime
The usage of SpyTime structure is that,
when we call (*ConfigurableSleeper).Sleep(),
it will called the function value stored in ConfigurableSleeper struct field.
For example: in main(), we set (*ConfigurableSleeper).sleep(time.Duration)
to time.Sleep(), to actually sleep for seconds, and in test, we set
(*ConfigurableSleeper).sleep(time.Duration)
to (*SpyTime).Sleep(time.Duration), which just record the slept duration
to a field of SpyTime, make us check the slept time easily.