Use sync.Pool to make buf act like global obj
sync.Pool is a feature in Go that allows you to amortize allocation cost.
With our current code, when we allocate the buffers,
we do so for the lifetime of the foo function.
But what if we were able to allocate it once, store it in a global variable,
then just reuse it infinitely (resetting the contents every time while keeping the capacity)?
Doing this manually would be a little complicated and also wouldn’t work concurrently.
That’s what sync.Pool is for. sync.Pool allows you to Get an already allocated object
and Put it back when you are done with it.
If there are no objects availabile in the pool and you request one,
it will call the New function you define to allocate one.
Let’s create a pool for the hash object and one for the hash sum (256 byte slice).

Ref:
    - [https://chris124567.github.io/2021-06-21-go-performance/]